+++ src/components/global/StateSelect.jsx
@@ -10,7 +10,7 @@ export default function StateSelect({ id, name, value, onChange, required, class
             onChange={onChange}
             required={required}
             className={cn(
-                "h-9 w-full min-w-0 rounded-4xl border border-input bg-input/30 px-3 py-1 text-base transition-colors outline-none placeholder:text-muted-foreground focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 md:text-sm appearance-none cursor-pointer",
+                "h-9 w-full min-w-0 rounded-4xl border border-input bg-input/30 px-3 py-1 text-base transition-colors outline-none placeholder:text-muted-foreground focus-visible:border-[var(--primary-color)] focus-visible:ring-[3px] focus-visible:ring-[var(--primary-color)]/50 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 md:text-sm appearance-none cursor-pointer",
                 className
             )}
         >